# api-contracts/automation-platform-scheduler-admission.openapi.yaml
openapi: 3.0.3
info:
  title: AutomationPlatform API
  version: 0.7.0
  description: |
    Contract-first description of the AutomationPlatform scheduler
    control-admission observability surface.

    This initial contract is derived from the frozen
    v0.25.0-phase-a25 runtime and intentionally covers only the
    admission status, bounded in-memory history, and restart-safe
    durable-history HTTP endpoints.

    Runtime Swagger/OpenAPI registration is intentionally out of scope
    for this version.
...
paths:
...
  /operations/scheduler/commands:

    get:
      tags:
        - Scheduler Operations
      operationId: getSchedulerCommandAuditHistory
      summary: Read recent scheduler control-command audit history
      parameters:
        - $ref: '#/components/parameters/SchedulerOperationalHistoryLimit'
      responses:
        '200':
          description: Recent scheduler control-command audit history.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerControlAuditHistory'
        '400':
          description: Invalid scheduler control-audit request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerControlAuditValidationError'
    post:
      tags:
        - Scheduler Operations

      operationId: executeSchedulerCommand

      summary: Execute a governed scheduler control command

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulerCommandRequest'

      responses:

        '200':
          description: Scheduler command executed successfully.

          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SchedulerControlResult'
                  - $ref: '#/components/schemas/SchedulerRestartResult'

        '400':
          description: Invalid scheduler control request.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerCommandValidationError'

        '409':
          description: Scheduler command was not executed due to governed conflict or admission denial.

          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SchedulerRestartResult'
                  - $ref: '#/components/schemas/SchedulerSupersededResponse'
                  - $ref: '#/components/schemas/SchedulerCommandAdmissionDeniedResponse'

        '500':
          description: Scheduler control command execution failed unexpectedly.

          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerCommandExecutionError'
  /operations/scheduler/start:
    post:
      tags:
        - Scheduler Operations
      operationId: executeSchedulerStart
      summary: Start the scheduler through governed coordinated control
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulerDedicatedCommandRequest'
      responses:
        '200':
          description: Scheduler start command executed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerControlResult'
        '400':
          description: Invalid scheduler control request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerCommandValidationError'
        '409':
          description: Scheduler start was not executed because of governed conflict or admission denial.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SchedulerRestartResult'
                  - $ref: '#/components/schemas/SchedulerSupersededResponse'
                  - $ref: '#/components/schemas/SchedulerCommandAdmissionDeniedResponse'
        '500':
          description: Scheduler control execution failed unexpectedly.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerCommandExecutionError'
  /operations/scheduler/stop:
    post:
      tags:
        - Scheduler Operations
      operationId: executeSchedulerStop
      summary: Stop the scheduler through governed coordinated control
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulerDedicatedCommandRequest'
      responses:
        '200':
          description: Scheduler stop command executed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerControlResult'
        '400':
          description: Invalid scheduler control request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerCommandValidationError'
        '409':
          description: Scheduler stop was not executed because of governed conflict or admission denial.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SchedulerRestartResult'
                  - $ref: '#/components/schemas/SchedulerSupersededResponse'
                  - $ref: '#/components/schemas/SchedulerCommandAdmissionDeniedResponse'
        '500':
          description: Scheduler control execution failed unexpectedly.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerCommandExecutionError'
  /operations/scheduler/restart:
    post:
      tags:
        - Scheduler Operations
      operationId: executeSchedulerRestart
      summary: Restart the scheduler through governed coordinated recovery
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulerDedicatedCommandRequest'
      responses:
        '200':
          description: Scheduler restart command executed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerRestartResult'
        '400':
          description: Invalid scheduler control request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerCommandValidationError'
        '409':
          description: Scheduler restart was not executed because of governed conflict or admission denial.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/SchedulerRestartResult'
                  - $ref: '#/components/schemas/SchedulerSupersededResponse'
                  - $ref: '#/components/schemas/SchedulerCommandAdmissionDeniedResponse'
        '500':
          description: Scheduler control execution failed unexpectedly.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulerCommandExecutionError'
...
components:
  parameters:
    SchedulerOperationalHistoryLimit:
      name: limit
      in: query
      required: false
      description: Maximum number of recent operational audit records to return.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  schemas:
...
    SchedulerDedicatedCommandRequest:
      type: object
      description: Optional request metadata for a scheduler command encoded by the dedicated URL.
      properties:
        requestKey:
          type: string
          minLength: 1
          maxLength: 128
    SchedulerCommandRequest:
      type: object
      additionalProperties: false

      required:
        - command

      properties:

        command:
          $ref: '#/components/schemas/SchedulerAdmissionCommand'

        requestKey:
          type: string
          minLength: 1
          maxLength: 128
    SchedulerControlDisposition:
      type: string

      enum:
        - executed
        - rejected
    SchedulerRuntimeState:
      type: string

      enum:
        - idle
        - running
    SchedulerControlResult:
      type: object
      additionalProperties: false

      required:
        - command
        - disposition
        - previousState
        - currentState
        - changed
        - reason

      properties:

        command:
          type: string

          enum:
            - start
            - stop

        disposition:
          $ref: '#/components/schemas/SchedulerControlDisposition'

        previousState:
          $ref: '#/components/schemas/SchedulerRuntimeState'

        currentState:
          $ref: '#/components/schemas/SchedulerRuntimeState'

        changed:
          type: boolean

        reason:
          type: string
          nullable: true
    SchedulerRestartDisposition:
      type: string

      enum:
        - executed
        - rejected
    SchedulerRestartResult:
      type: object
      additionalProperties: false

      required:
        - command
        - disposition
        - previousGeneration
        - currentGeneration
        - previousState
        - currentState
        - changed
        - reason

      properties:

        command:
          type: string
          enum:
            - restart

        disposition:
          $ref: '#/components/schemas/SchedulerRestartDisposition'

        previousGeneration:
          type: integer

        currentGeneration:
          type: integer

        previousState:
          $ref: '#/components/schemas/SchedulerRuntimeState'

        currentState:
          $ref: '#/components/schemas/SchedulerRuntimeState'

        changed:
          type: boolean

        reason:
          type: string
          nullable: true
    SchedulerSupersededResponse:
      type: object
      additionalProperties: false

      required:
        - command
        - disposition
        - attemptedGeneration
        - observedGeneration
        - changed
        - reason

      properties:

        command:
          type: string
          enum:
            - restart

        disposition:
          type: string
          enum:
            - superseded

        attemptedGeneration:
          type: integer

        observedGeneration:
          type: integer

        changed:
          type: boolean
          enum:
            - false

        reason:
          type: string
          enum:
            - Superseded by a later durable scheduler generation.
    SchedulerCommandAdmissionDeniedResponse:
      type: object
      additionalProperties: false

      required:
        - kind
        - command
        - reason

      properties:

        kind:
          type: string
          enum:
            - admission_denied

        command:
          $ref: '#/components/schemas/SchedulerAdmissionCommand'

        reason:
          $ref: '#/components/schemas/SchedulerAdmissionDenialReason'
    SchedulerCommandValidationError:
      type: object
      additionalProperties: false

      required:
        - error
        - message

      properties:

        error:
          type: string
          enum:
            - invalid_scheduler_control_request

        message:
          type: string
    SchedulerCommandExecutionError:
      type: object
      additionalProperties: false

      required:
        - error
        - message

      properties:

        error:
          type: string
          enum:
            - scheduler_control_error

        message:
          type: string
          enum:
            - Unable to execute scheduler control command.
    SchedulerAdmissionCommand:
      type: string

      enum:
        - start
        - stop
        - restart
    SchedulerAdmissionDenialReason:
      type: string

      enum:
        - scheduler_standby
        - scheduler_fail_closed
        - scheduler_stopped
...
    SchedulerStatusRuntimeState:
      type: string
      enum:
        - idle
        - running
        - stopped
        - failed
...
    SchedulerControlAuditCommand:
      type: string
      enum:
        - start
        - stop
    SchedulerControlAuditDisposition:
      type: string
      enum:
        - executed
        - noop
        - rejected
    SchedulerControlAuditStatus:
      type: string
      enum:
        - pending
        - completed
        - failed
    SchedulerControlAuditItem:
      type: object
      additionalProperties: false
      required:
        - auditId
        - publicId
        - requestKey
        - command
        - auditStatus
        - disposition
        - previousState
        - currentState
        - changed
        - reason
        - errorMessage
        - createdAtUtc
        - completedAtUtc
      properties:
        auditId:
          type: string
          description: Decimal string representation of the persisted bigint audit identifier.
        publicId:
          type: string
        requestKey:
          type: string
          nullable: true
        command:
          $ref: '#/components/schemas/SchedulerControlAuditCommand'
        auditStatus:
          $ref: '#/components/schemas/SchedulerControlAuditStatus'
        disposition:
          type: string
          nullable: true
          allOf:
            - $ref: '#/components/schemas/SchedulerControlAuditDisposition'
        previousState:
          type: string
          nullable: true
          allOf:
            - $ref: '#/components/schemas/SchedulerStatusRuntimeState'
        currentState:
          type: string
          nullable: true
          allOf:
            - $ref: '#/components/schemas/SchedulerStatusRuntimeState'
        changed:
          type: boolean
          nullable: true
        reason:
          type: string
          nullable: true
        errorMessage:
          type: string
          nullable: true
        createdAtUtc:
          type: string
          format: date-time
        completedAtUtc:
          type: string
          format: date-time
          nullable: true
    SchedulerControlAuditHistory:
      type: object
      additionalProperties: false
      required:
        - count
        - items
      properties:
        count:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/SchedulerControlAuditItem'
    SchedulerControlAuditValidationError:
      type: object
      additionalProperties: false
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - invalid_scheduler_control_audit_request
        message:
          type: string
          enum:
            - limit must be an integer from 1 to 100.
...